#4, removed GameStateManager and changed folder structure.
