+++ wordleBattleProject/core/src/com/mygdx/game/controller/Controller.java
@@ -1,7 +1,7 @@
 package com.mygdx.game.controller;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.mygdx.game.model.State;
+import com.mygdx.game.model.states.State;
 import com.mygdx.game.view.View;
 
 /**
