+++ wordleBattleProject/core/src/com/mygdx/game/controller/MainMenuController.java
@@ -1,8 +1,6 @@
 package com.mygdx.game.controller;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.mygdx.game.model.State;
-import com.mygdx.game.view.View;
 
 /**
  * @author Marcus Birkeland
