+++ wordleBattleProject/core/src/com/mygdx/game/model/states/GameSetupState.java
@@ -1,0 +1,18 @@
+package com.mygdx.game.model.states;
+
+public class GameSetupState extends State {
+    private static GameSetupState instance = new GameSetupState();
+
+    private GameSetupState() {
+    }
+
+    public static GameSetupState getInstance() {
+        return instance;
+    }
+
+    @Override
+    public void update(float dt) {
+
+    }
+
+}
