+++ wordleBattleProject/core/src/com/mygdx/game/model/states/HighScoreState.java
@@ -1,0 +1,21 @@
+package com.mygdx.game.model.states;
+
+public class HighScoreState extends State {
+    private int[] highscores;
+    private static HighScoreState instance = new HighScoreState();
+
+    public HighScoreState() {
+
+    }
+
+    public static HighScoreState getInstance() {
+        return instance;
+    }
+
+    @Override
+    public void update(float dt) {
+
+    }
+
+
+}
