+++ wordleBattleProject/core/src/com/mygdx/game/model/states/TutorialState.java
@@ -1,0 +1,17 @@
+package com.mygdx.game.model.states;
+
+public class TutorialState extends State{
+
+
+    private static TutorialState instance = new TutorialState();
+
+    public TutorialState() {
+    }
+
+
+    @Override
+    public void update(float dt) {
+
+    }
+
+}
