--- wordleBattleProject/core/src/com/mygdx/game/states/State.java
@@ -1,19 +1,0 @@
-package com.mygdx.game.states;
-
-import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.math.Vector3;
-
-public abstract class State {
-    protected OrthographicCamera cam;
-    protected Vector3 mouse;
-    protected GameStateManager gsm;
-
-    public State(GameStateManager gsm) {
-        this.gsm = gsm;
-        cam = new OrthographicCamera();
-        mouse = new Vector3();
-    }
-
-    public abstract void update(float dt);
-}
